3.3_higher-order_functions.ipynb: small fixes (#99)
Jupyter Notebook seems to have a problem parsing and rendering the
italic fonts here.

According to the [chisel3 API reference][1], PriorityMux() should
accept a Seq of (Bool, T) where T <: Data, but in the exercise
description the two arguments in a tuple are mistakenly swapped.
Fixing the confusing description. The solution provided is correct
though.

[1]: https://www.chisel-lang.org/api/latest/chisel3/util/PriorityMux$.html#apply[T%3C:chisel3.Data](in:Seq[(chisel3.Bool,T)]):T

The solution for the exercise "Decoupled Arbiter" still contains a
for loop. Since this chapter is introducing functional programming
concepts, let's eliminate this for loop too.